refactor: consolidate root level test files into tests folder
