+++ package.json
@@ -12,7 +12,7 @@
     "clean": "rm -rf dist",
     "lint": "tsc --noEmit",
     "scrape": "tsx scrape-cli.ts",
-    "test-mongo": "tsx test-mongo.ts",
+    "test-mongo": "tsx tests/test-mongo.ts",
     "test-dnl": "tsx src/services/dnl/validationTest.ts"
   },
   "dependencies": {
