+++ tests/test-rbac.ts
@@ -1,3 +1,4 @@
+// @ts-ignore
 import jwt from "jsonwebtoken";
 import dotenv from "dotenv";
 dotenv.config();
