+++ tests/test-static.js
@@ -1,4 +1,5 @@
 const express = require('express');
+const path = require('path');
 const app = express();
-app.use(express.static('public'));
+app.use(express.static(path.join(__dirname, '../public')));
 app.listen(3001, () => console.log('started'));
